Refactor
Разбил на отдельные файлы (мне пока не удобно так работать) убрал некоторые предупреждения
