+++ internal/service_consts/consts.go
@@ -1,0 +1,4 @@
+package service_consts
+
+const Education = "education"
+const Solution = "solution"
